order model (passing tests)
n.b. removed the validate_presence_of expectations on order as they don't
expect a default value to be substituted in (e.g. order_date is nil and
should stay nil). The validation stays in place as a safety measure,
though the before_validation hooks should ensure that the data is always
valid.